+++ R/Gibbs_OMFA.R
@@ -63,10 +63,10 @@
     }
     psi.beta       <- switch(uni.prior,  isotropic=psi.beta[which.max(.ndeci(psi.beta))], psi.beta)
     pi.prop        <- c(cluster$pi.prop, vector("integer", G - length(cluster$pi.prop)))
-    mu             <- cbind(mu,  vapply(seq_len(G - length(cluster$pi.prop)), function(g) .sim_mu_p(P=P, sig.mu.sqrt=sig.mu.sqrt, mu.zero=mu.zero), numeric(P)))
+    mu             <- cbind(mu, replicate(G - length(cluster$pi.prop), .sim_mu_p(P=P, sig.mu.sqrt=sig.mu.sqrt, mu.zero=mu.zero), simplify="array"))
     eta            <- .sim_eta_p(N=N, Q=Q)
     lmat           <- if(Q0) array(vapply(Gseq, function(g) .sim_load_p(Q=Q, P=P, sigma.l=sigma.l), numeric(P * Q)), dim=c(P, Q, G)) else array(0, dim=c(P, 0, G))
-    psi.inv        <- vapply(Gseq, function(g) .sim_psi_ip(P=P, psi.alpha=psi.alpha, psi.beta=psi.beta), numeric(P))
+    psi.inv        <- replicate(G, .sim_psi_ip(P=P, psi.alpha=psi.alpha, psi.beta=psi.beta), simplify="array")
     psi.inv        <- if(uni) t(psi.inv) else psi.inv
     if(Q0 &&   Q    < min(N - 1, Ledermann(P))) {
       for(g in which(nn     > P))   {